feature: add model registry to allow centralized access of the supported model and default model
